First commit. All changes are made from original timetaggerd.
win64/Makefile
--------------
@cp prequesits/vcredist_x86.exe dist/system ->
	@cp prequesits/vcredist_x64.exe dist/system
@cp ../win32/prequesits/Driver/i386/WdfCoInstaller01005.dll dist/system/i386 ->
	@cp ../win32/prequesits/Driver/i386/WdfCoInstaller01009.dll dist/system/i386
@cp ../win32/prequesits/Driver/i386/WinUSBCoInstaller.dll dist/system/i386 ->
 	@cp ../win32/prequesits/Driver/i386/WinUSBCoInstaller2.dll dist/system/i386
@cp prequesits/Driver/amd64/WdfCoInstaller01005.dll dist/system/amd64 ->
 	@cp prequesits/Driver/amd64/WdfCoInstaller01009.dll dist/system/amd64
@cp prequesits/Driver/amd64/WinUSBCoInstaller.dll dist/system/amd64 ->
	@cp prequesits/Driver/amd64/WinUSBCoInstaller2.dll dist/system/amd64

Folder `timetaggerd` was removed from `dist-clean` as it is causing builds
to fail. Cause appears to be related to the static linking of libraries in
`timetaggerd/httpd`, possibly because they were compiled on another system
and will have to recompiled to run correctly.

win64/build
-----------
vcredist_x86.exe was removed

Makefile
--------
Added line to `make all` for compilation of `setup/timetagger.nsi`