Reverted to pre-make script state. Python 3.5/3.6 compilation supported.
The make script was found to be defective and removed. Compilation against
Python 3.5/3.6 is supported, but unfortunately the graphical backend for
pi3diamond does not appear to support Python 3. This may or may not be resolved
in the future, however support for compilation against Python 2 and Python 3.5/3.6
will be maintained in the timetagger.